requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pydantic>=2.7.0
pydantic-ai>=0.0.14
anthropic>=0.39.0
supabase>=2.9.1
graphiti-core>=0.3.0
neo4j>=5.26.0
google-api-python-client>=2.154.0
google-auth>=2.36.0
google-auth-oauthlib>=1.2.1
google-auth-httplib2>=0.2.0
youtube-transcript-api>=0.6.2
openai>=1.54.3
PyJWT>=2.9.0
python-jose[cryptography]>=3.3.0
cryptography>=41.0.0
python-multipart>=0.0.6
httpx>=0.26.0
aiofiles>=23.2.0
PyPDF2>=3.0.0
python-docx>=1.1.0
pdfplumber>=0.11.0
pandas>=2.2.0
openpyxl>=3.1.0
python-dotenv>=1.0.0
